pyproject.toml
[tool.poetry]
name = "json-fastapi"
version = "0.1.0"
description = "Library to add basic read-only JSON APIs to FastAPI projects"
authors = ["Tom Stockwell <2060486+stocky37@users.noreply.github.com>"]
packages = [
    { include = "json_fastapi", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {version = "^0.55.1", optional = true}
tinydb = {version = "^4.1.1", optional = true}
python-slugify = "^4.0.0"
linkheader = "^0.4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pre-commit = "^2.4.0"
black = "^19.10b0"

[tool.poetry.extras]
all = ["fastapi", "tinydb"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
